+++ apps/matheportal-shell-ui/public/federation.manifest.json
