+++ apps/matheportal-shell-ui/src/app/app.component.ts
