+++ apps/matheportal-shell-ui/src/bootstrap.ts
