+++ apps/matheportal-shell-ui/src/styles.scss
