--- apps/portal-shell/federation.config.js
@@ -1,28 +1,0 @@
-const { withNativeFederation, shareAll } = require('@angular-architects/native-federation/config');
-
-module.exports = withNativeFederation({
-  name: 'portal-shell',
-
-
-  shared: {
-    ...shareAll({ singleton: true, strictVersion: true, requiredVersion: 'auto' }),
-  },
-
-  skip: [
-    'rxjs/ajax',
-    'rxjs/fetch',
-    'rxjs/testing',
-    'rxjs/webSocket',
-    // Add further packages you don't need at runtime
-  ],
-
-  // Please read our FAQ about sharing libs:
-  // https://shorturl.at/jmzH0
-
-  features: {
-    // New feature for more performance and avoiding
-    // issues with node libs. Comment this out to
-    // get the traditional behavior:
-    ignoreUnusedDeps: true
-  }
-});
